+++ src/components-v2/common/dialogs/SettDialogTitle.tsx
@@ -36,7 +36,7 @@ export const SettDialogTitle = ({ sett, mode }: Props): JSX.Element => {
 				<Grid item className={classes.logoContainer}>
 					<img
 						className={classes.settLogo}
-						src={`/assets/icons/${sett.asset.toLowerCase()}.png`}
+						src={`/assets/icons/${sett.vaultAsset.toLowerCase()}.png`}
 						alt={`Badger ${sett.name} Vault Symbol`}
 					/>
 				</Grid>