+++ src/components-v2/landing/SettListItem.tsx
@@ -69,7 +69,7 @@ const useStyles = makeStyles((theme) => ({
 
 export interface SettListItemProps {
 	sett: Sett;
-	balance?: string;
+	balance?: BigNumber;
 	balanceValue?: string;
 	accountView?: boolean;
 	currency: Currency;
@@ -77,7 +77,7 @@ export interface SettListItemProps {
 }
 
 const SettListItem = observer(
-	({ sett, balanceValue, currency, period, accountView = false }: SettListItemProps): JSX.Element => {
+	({ sett, balance, balanceValue, currency, period, accountView = false }: SettListItemProps): JSX.Element => {
 		const { user, network, router, wallet } = useContext(StoreContext);
 		const [openDepositDialog, setOpenDepositDialog] = useState(false);
 		const [openWithdrawDialog, setOpenWithdrawDialog] = useState(false);
@@ -90,8 +90,9 @@ const SettListItem = observer(
 		const displayValue = balanceValue ? balanceValue : inCurrency(new BigNumber(sett.value), currency);
 		const multiplier = !sett.deprecated ? user.accountDetails?.multipliers[sett.vaultToken] : undefined;
 
-		const canWithdraw = user.getSettBalance(sett).balance > 0;
-		const isDisabled = sett.hasBouncer && !user.permissions?.viewSettShop;
+		const canWithdraw = balance ? balance.gt(0) : false;
+		// sett is disabled if they are internal setts, or have a bouncer and use has no access
+		const isDisabled = !user.onGuestList(sett);
 
 		const goToSettDetail = async () => {
 			await router.goTo(routes.settDetails, { settName: sett.slug, accountView });
@@ -126,7 +127,7 @@ const SettListItem = observer(
 					</Grid>
 					<Grid item xs={12} md className={classes.nonClickableSection}>
 						<SettActionButtons
-							isWithdrawDisabled={!canWithdraw || !wallet.connectedAddress || isDisabled}
+							isWithdrawDisabled={!wallet.connectedAddress || !canWithdraw}
 							isDepositDisabled={!wallet.connectedAddress || isDisabled}
 							onWithdrawClick={() => setOpenWithdrawDialog(true)}
 							onDepositClick={() => setOpenDepositDialog(true)}