+++ src/components-v2/sett-detail/holdings/Holdings.tsx
@@ -31,58 +31,57 @@ interface Props {
 	settBalance: SettBalance;
 }
 
-export const Holdings = observer(
-	({ tokenBalance, settBalance, sett, badgerSett }: Props): JSX.Element => {
-		const { setts } = React.useContext(StoreContext);
-		const isMediumSizeScreen = useMediaQuery(useTheme().breakpoints.up('sm'));
-		const classes = useStyles();
+export const Holdings = observer(({ tokenBalance, settBalance, sett, badgerSett }: Props): JSX.Element | null => {
+	const { setts, user } = React.useContext(StoreContext);
+	const isMediumSizeScreen = useMediaQuery(useTheme().breakpoints.up('sm'));
+	const classes = useStyles();
+	const canDeposit = user.onGuestList(sett);
 
-		if (!hasBalance(settBalance) && !TokenBalance.hasBalance(tokenBalance)) {
-			return (
-				<Grid container>
-					<NoHoldings sett={sett} badgerSett={badgerSett} />
-				</Grid>
-			);
-		}
-
-		const { earnedBalance, earnedValue, balance, value } = settBalance;
-		const logo = `/assets/icons/${sett.asset.toLowerCase()}.png`;
-
-		const depositToken = setts.getToken(sett.underlyingToken);
-		const decimals = depositToken?.decimals || 18;
-
+	if (!hasBalance(settBalance) && !TokenBalance.hasBalance(tokenBalance)) {
 		return (
 			<Grid container>
-				<Grid container>
-					<Typography className={classes.settInfoTitle}>Your Vault Info</Typography>
-				</Grid>
-				<Grid container spacing={1}>
-					<Grid item xs={12} sm>
-						<HoldingItem
-							name="Total Balance"
-							logo={logo}
-							balance={balance}
-							value={value}
-							decimals={decimals}
-							helpIcon={<TokenDistributionIcon settBalance={settBalance} />}
-						/>
-					</Grid>
-					<Grid item xs={12} sm>
-						<HoldingItem
-							name="Total Earned"
-							logo={logo}
-							balance={earnedBalance}
-							value={earnedValue}
-							decimals={decimals}
-						/>
-					</Grid>
-					{isMediumSizeScreen && (
-						<Grid item xs={12} sm>
-							<HoldingsActionButtons />
-						</Grid>
-					)}
-				</Grid>
+				<NoHoldings sett={sett} badgerSett={badgerSett} />
 			</Grid>
 		);
-	},
-);
+	}
+
+	const { earnedBalance, earnedValue, balance, value } = settBalance;
+	const logo = `/assets/icons/${sett.vaultAsset.toLowerCase()}.png`;
+
+	const depositToken = setts.getToken(sett.underlyingToken);
+	const decimals = depositToken?.decimals || 18;
+
+	return (
+		<Grid container>
+			<Grid container>
+				<Typography className={classes.settInfoTitle}>Your Vault Info</Typography>
+			</Grid>
+			<Grid container spacing={1}>
+				<Grid item xs={12} sm>
+					<HoldingItem
+						name="Total Balance"
+						logo={logo}
+						balance={balance}
+						value={value}
+						decimals={decimals}
+						helpIcon={<TokenDistributionIcon settBalance={settBalance} />}
+					/>
+				</Grid>
+				<Grid item xs={12} sm>
+					<HoldingItem
+						name="Total Earned"
+						logo={logo}
+						balance={earnedBalance}
+						value={earnedValue}
+						decimals={decimals}
+					/>
+				</Grid>
+				{isMediumSizeScreen && (
+					<Grid item xs={12} sm>
+						<HoldingsActionButtons canDeposit={canDeposit} />
+					</Grid>
+				)}
+			</Grid>
+		</Grid>
+	);
+});