+++ src/tests/sett-details/SpecsCard.test.tsx
@@ -1,6 +1,6 @@
+import SpecsCard from 'components-v2/sett-detail/specs/SpecsCard';
 import React from 'react';
 import { checkSnapshot } from 'tests/utils/snapshots';
-import { SpecsCard } from '../../components-v2/sett-detail/specs/SpecsCard';
 import { SAMPLE_BADGER_SETT, SAMPLE_SETT } from '../utils/samples';
 
 describe('Specs Section', () => {